// apps/gateway.old/service/src/auth/cerbos/permission.enum.ts
export enum CERBOS_ACTIONS {
    CREATE = 'create',
    READ = 'read',
    UPDATE = 'update',
    DELETE = 'delete',
    LIST = 'list',
    SEARCH = 'search',
    ASSIGN = 'assign',
    UNASSIGN = 'unassign',
    APPROVE = 'approve',
    REJECT = 'reject',
    PUBLISH = 'publish',
    UNPUBLISH = 'unpublish',
    EXPORT = 'export',
    IMPORT = 'import',
    SEND = 'send',
    RECEIVE = 'receive',
    NOTIFY = 'notify',
    GENERATE = 'generate',
    STOP = 'stop',
    VERIFY = 'verify',
    RETRY = 'retry'
}

export enum CERBOS_RESOURCES {
    STORY = 'story',
    COMMENT = 'comment',
    TRANSLATION = 'translation',
    IVRR_FILES = 'ivrr-files',
    SOCIAL_MESSAGE = 'social-message',
    USER = 'user',
    REJECT_REASON = 'reject-reason',
    MODERATOR = 'moderator',
    ORGANISATION = 'organisation'
}